requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
httpx
aiogram
pydantic
